refactor: Retrofit 관련 코드 Hilt로 리팩터렁
issue: #41
