+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/app/DdaraogaeApplication.kt
@@ -1,0 +1,8 @@
+package com.nbcfinalteam2.ddaraogae.app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class DdaraogaeApplication: Application() {
+}
